pyproject.toml
[project]
name = "submonitorapi"
version = "0.1.0"
description = ""
authors = [
    {name = "Aleksei-Kutuzov",email = "alekseikutuzov2@yandex.com"}
]
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.135.1,<0.136.0)",
    "starlette (>=0.52.1,<0.53.0)",
    "sqlalchemy (>=2.0.48,<3.0.0)",
    "python-jose (>=3.5.0,<4.0.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)",
    "alembic (>=1.18.4,<2.0.0)",
    "passlib (==1.7.4)",
    "bcrypt (==4.1.2)",
    "pydantic[email] (>=2.12.5,<3.0.0)",
    "python-multipart (>=0.0.22,<0.0.23)",
    "python-dateutil (>=2.9.0.post0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
